+++ R/helpers.R
@@ -40,16 +40,6 @@ gsub_warn_on_error <- function(pattern, replacement, x, ignore.case = FALSE, per
            })
 }
 
-# # finds start and end of a match and returns that substring
-# find_match <- function(x, keep, fixed) {
-#   lapply(x, function(val) {
-#     match_found <- regexpr(pattern = keep, text = val, fixed = fixed)
-#     matched <- list(start = as.integer(match_found),
-#                     end = -1 + as.integer(match_found) + as.integer(attributes(match_found)['match.length']))
-#     substr(val, matched$start, matched$end)
-#   })
-# }
-
 # works exactly like round(), but rounds `round(44.55, 1)` as 44.6 instead of 44.5
 # and adds decimal zeroes until `digits` is reached when force_zero = TRUE
 round2 <- function(x, digits = 0, force_zero = TRUE) {
@@ -57,10 +47,19 @@ round2 <- function(x, digits = 0, force_zero = TRUE) {
   # https://stackoverflow.com/a/12688836/4575331
   val <- (trunc((abs(x) * 10 ^ digits) + 0.5) / 10 ^ digits) * sign(x)
   if (digits > 0 & force_zero == TRUE) {
-    val[val != as.integer(val) & !is.na(val)] <- paste0(val[val != as.integer(val) & !is.na(val)],
-                                                        strrep("0", max(0, digits - nchar(gsub(".*[.](.*)$", "\\1", val[val != as.integer(val) & !is.na(val)])))))
+    values_trans <- val[val != as.integer(val) & !is.na(val)]
+    val[val != as.integer(val) & !is.na(val)] <- paste0(values_trans,
+                                                        strrep("0", 
+                                                               max(0, 
+                                                                   digits - nchar(
+                                                                     format(
+                                                                       as.double(
+                                                                         gsub(".*[.](.*)$", 
+                                                                              "\\1",
+                                                                              values_trans)),
+                                                                       scientific = FALSE)))))
   }
-  val
+  as.double(val)
 }
 
 getdecimalplaces <- function(x, minimum = 0, maximum = 3) {
@@ -70,7 +69,7 @@ getdecimalplaces <- function(x, minimum = 0, maximum = 3) {
   if (minimum > maximum) {
     minimum <- maximum
   }
-  max_places <- max(unlist(lapply(strsplit(sub('0+$', '', 
+  max_places <- max(unlist(lapply(strsplit(sub("0+$", "", 
                                                as.character(x * 100)), ".", fixed = TRUE),
                                   function(y) ifelse(length(y) == 2, nchar(y[2]), 0))), na.rm = TRUE)
   max(min(max_places,
@@ -90,4 +89,26 @@ cqv <- function(x, na.rm = TRUE) {
   # m = p. p[k] = k / (n + 1). Thus p[k] = E[F(x[k])]. This is used by Minitab and by SPSS.
   quartiles <- stats::quantile(x, probs = c(0.25, 0.75), na.rm = na.rm, type = 6)
   (quartiles[2] - quartiles[1]) / (quartiles[2] + quartiles[1])
+}
+
+# on checks if the input is a valid date type. It supports all date types, including \code{Date}, \code{POSIXct} and \code{POSIXlt}.
+is.Date <- function(x) {
+  inherits(x, c("Date", "POSIXt"))
+}
+
+year <- function(x) {
+  as.POSIXlt(x, tz = tz(x))$year + 1900
+}
+
+tz <- function(x) {
+  # from pkg lubridate
+  tzone <- attr(x, "tzone")[[1]]
+  if (is.null(tzone) && !inherits(x, "POSIXt")) 
+    return("UTC")
+  if (is.character(tzone) && nzchar(tzone)) 
+    return(tzone)
+  tzone <- attr(as.POSIXlt(x[1]), "tzone")[[1]]
+  if (is.null(tzone)) 
+    return("UTC")
+  tzone
 }